Organize system route workflow folders
